feat(config): Add --print-config option to print configuration and exit
Added a new option to print the configuration and exit. This includes a
new `print_config` function and an update to the `main` function to
include the `--print-config` option. Additionally, minor
formatting/typo improvements were made to the `Config` class and
`PucotiConfig` class.


feat(config): Add --print-config option to print configuration and exit

Added a new option to print the configuration and exit. This includes a
new `print_config` function and an update to the `main` function to
include the `--print-config` option. Additionally, minor
formatting/typo improvements were made to the `Config` class and
`PucotiConfig` class.