fix: Always reset the prupose when app opens
Add a `force` parameter to the `set_purpose` method in `Context` class
to allow overriding the current purpose. Update `MainScreen` to use the
`force` parameter when setting an empty purpose on initialization.


fix: Always reset the prupose when app opens

Add a `force` parameter to the `set_purpose` method in `Context` class
to allow overriding the current purpose. Update `MainScreen` to use the
`force` parameter when setting an empty purpose on initialization.