fix(config): Handle missing fields in merge method
Fix issue where missing fields in the configuration were not being set to their default values. Now, if a field is not present in the input values, it is set to its current value from the instance.


fix(config): Handle missing fields in merge method

Fix issue where missing fields in the configuration were not being set to their default values. Now, if a field is not present in the input values, it is set to its current value from the instance.