feat(config): Validate existence of config file before loading
Add validation for the existence of the configuration file before loading it. If the configuration file does not exist, raise a `typer.BadParameter` error to inform the user.


feat(config): Validate existence of config file before loading

Add validation for the existence of the configuration file before loading it. If the configuration file does not exist, raise a `typer.BadParameter` error to inform the user.